AK: Remove whitespace skipping feature from AK's Base64 decoder
This was added in commit e5c00222622956bf6b67702da100f1cd2f71a9e2 as a
partial implementation of what is now LibWeb's forgiving Base64 decoder.
All use cases within LibWeb that require whitespace skipping now use
that implementation instead.

Removing this feature from AK allows us to know the exact output size of
a decoded Base64 string. We can still trim whitespace at the start and
end of the input though; for example, this is useful when reading from a
file that may have a newline at the end of the file.