RequestServer: Avoid race condition between timeout and socket creation
Part of this issue was fixed in f285f81d3c0c713ed2435545af96ff9551cd8f59
but that only addressed the first layer of deferred_invoke, ignoring the
second one (which would cause a race if a request was sent to a host
immediately following a timeout event from the same host).
Fixes #23840.