Revert "Kernel+SystemServer: Make KCOVDevice a character device"
This reverts commit 2f35935e27f6eb5284d3bb490ffd9ec789bc7321.

For KCOV to be performant (or at least not even slower) we need to
mmap the PC buffer from both user and kernel space at the same time.
You can't mmap a character device, so this change didn't make sense.

Plus even if we did invent a new method to exfiltrate the coverage
information out of the kernel, it would be incompatible with existing
kernel fuzzers. That would be kind of annoying. 🙃