Utilities/mount: Resolve regression on mounting non-storage-backed FSes
Without this patch, we fail on manually mounting RAMFS (which I tested
for) but any filesystem that is not backed by actual storage will fail.
This bug was introduced in 1d889264aa10a7d00bf38331df15f2d2c76f0d82 and
now is resolved by checking if the source fd is negative, to avoid fail
of the fstat call on it.