Revert "LibCore: Add Core::deferred_invoke_if(F, Condition)"
This reverts commit 4fa5b48e192612db39ce0588e5bd0351a1dc6cb4.

The commit tried to add an unused function that continued our tradition
of not properly waiting for things but instead flooding event loop with
condition checks that delay firing of the event. I think this is a
fundamentally flawed approach. See also checks for
`fire_when_not_visible` in LibCore/EventLoopImplementationUnix.cpp.